QnA_Bot/requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv